requirements.txt
streamlit==1.12.2
opencv-python==4.6.0.66
scikit-image==0.19.3
asyncio==3.4.3
pdfplumber==0.8.0
torch==2.0.1
torchvision==0.15.2
pdf2image==1.16.2
Pillow==9.5.0
